FICGen/requirement.txt
transformers==4.27.4
accelerate==0.20.3
torch==1.12.1
torchvision==0.13.1
pillow==9.4.0
safetensors==0.5.3
opencv-python==4.10.0.84
diffusers==0.21.0
omegaconf
torch-fidelity==0.3.0
datasets==3.1.0
imagesize==1.4.1
matplotlib
numpy==1.21.6
numba==0.53.0
huggingface-hub==0.24.6
